Move NewExporter function to exporter.go
