+++ collector/collector.go
@@ -2,12 +2,9 @@ package collector
 
 import (
 	"strings"
-	"sync"
 	"time"
 
-	"github.com/go-kit/kit/log"
 	"github.com/go-kit/kit/log/level"
-	"github.com/peimanja/artifactory_exporter/config"
 	"github.com/prometheus/client_golang/prometheus"
 )
 
@@ -70,49 +67,6 @@ var (
 	artifactoryUp = newMetric("up", "", "Was the last scrape of Artifactory successful.", nil)
 )
 
-// Exporter collects JFrog Artifactory stats from the given URI and
-// exports them using the prometheus metrics package.
-type Exporter struct {
-	URI        string
-	cred       config.Credentials
-	authMethod string
-	sslVerify  bool
-	timeout    time.Duration
-	mutex      sync.RWMutex
-
-	up                              prometheus.Gauge
-	totalScrapes, jsonParseFailures prometheus.Counter
-	logger                          log.Logger
-}
-
-// NewExporter returns an initialized Exporter.
-func NewExporter(conf *config.Config) (*Exporter, error) {
-
-	return &Exporter{
-		URI:        conf.ArtiScrapeURI,
-		cred:       *conf.Credentials,
-		authMethod: conf.Credentials.AuthMethod,
-		sslVerify:  conf.ArtiSSLVerify,
-		timeout:    conf.ArtiTimeout,
-		up: prometheus.NewGauge(prometheus.GaugeOpts{
-			Namespace: namespace,
-			Name:      "up",
-			Help:      "Was the last scrape of artifactory successful.",
-		}),
-		totalScrapes: prometheus.NewCounter(prometheus.CounterOpts{
-			Namespace: namespace,
-			Name:      "exporter_total_scrapes",
-			Help:      "Current total artifactory scrapes.",
-		}),
-		jsonParseFailures: prometheus.NewCounter(prometheus.CounterOpts{
-			Namespace: namespace,
-			Name:      "exporter_json_parse_failures",
-			Help:      "Number of errors while parsing Json.",
-		}),
-		logger: conf.Logger,
-	}, nil
-}
-
 // Describe describes all the metrics ever exported by the Artifactory exporter. It
 // implements prometheus.Collector.
 func (e *Exporter) Describe(ch chan<- *prometheus.Desc) {
@@ -136,7 +90,7 @@ func (e *Exporter) Describe(ch chan<- *prometheus.Desc) {
 	ch <- e.jsonParseFailures.Desc()
 }
 
-// Collect fetches the stats from  Artifactiry and delivers them
+// Collect fetches the stats from  Artifactory and delivers them
 // as Prometheus metrics. It implements prometheus.Collector.
 func (e *Exporter) Collect(ch chan<- prometheus.Metric) {
 	e.mutex.Lock() // To protect metrics from concurrent collects.
